Palette - "Create Node" and "Create Label" buttons are working.
EventEmitter3 was replaced with PEvents (priority and promise-based eventemitter is required).

Take class 'no-drag' as negative conditional for the drag function.

Add connections by drag and drop (started): Draws an line as guide, remains to insert the link in the model and draw the real link.

Minor style refactoring

Add connections by drag and drop (progress): Insert the link in the model and draw the line.

Add connections by drag and drop (progress): Fix the arrow positions